AuthRequest applied on Shipping Company Resource

+++ src/shipping-company/shipping-company.service.spec.ts
@@ -10,6 +10,7 @@ import { ShippingCompanyService } from './shipping-company.service';
 
 /* Entity */
 import { ShippingCompany } from './entities/shipping-company.entity';
+import { User } from '@user/entities/user.entity';
 
 /* DTO's */
 import { UpdateShippingCompanyDto } from './dto/update-shipping-company.dto';
@@ -154,23 +155,23 @@ describe('ShippingCompanyService', () => {
   describe('create shipping companies services', () => {
     it('create should return a Shipping Company', async () => {
       const paymentMethod = generateShippingCompany();
-
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'create').mockReturnValue(paymentMethod);
       jest.spyOn(repository, 'save').mockResolvedValue(paymentMethod);
 
-      const { statusCode, data } = await service.create(paymentMethod);
+      const { statusCode, data } = await service.create(paymentMethod, userId);
       expect(statusCode).toBe(201);
       expect(data).toEqual(paymentMethod);
     });
 
     it('create should return Conflict Exception when name Shipping Company exists', async () => {
       const mock = generateShippingCompany();
-
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'create').mockReturnValue(mock);
       jest.spyOn(repository, 'save').mockResolvedValue(mock);
 
       try {
-        await service.create(mock);
+        await service.create(mock, userId);
       } catch (error) {
         expect(error).toBeInstanceOf(ConflictException);
         expect(error.message).toBe(
@@ -184,13 +185,14 @@ describe('ShippingCompanyService', () => {
     it('update should return message: have been modified', async () => {
       const mock = generateShippingCompany();
       const id = mock.id;
+      const userId: User['id'] = 1;
       const changes: UpdateShippingCompanyDto = { name: 'newName' };
 
       jest.spyOn(repository, 'findOne').mockResolvedValue(mock);
       jest.spyOn(repository, 'merge').mockReturnValue({ ...mock, ...changes });
       jest.spyOn(repository, 'save').mockResolvedValue(mock);
 
-      const { statusCode, message } = await service.update(id, changes);
+      const { statusCode, message } = await service.update(id, userId, changes);
       expect(repository.findOne).toHaveBeenCalledTimes(1);
       expect(repository.merge).toHaveBeenCalledTimes(1);
       expect(repository.save).toHaveBeenCalledTimes(1);
@@ -202,10 +204,11 @@ describe('ShippingCompanyService', () => {
 
     it('update should throw NotFoundException if Shipping Company does not exist', async () => {
       const id = 1;
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'findOne').mockResolvedValue(null);
 
       try {
-        await service.update(id, { name: 'newName' });
+        await service.update(id, userId, { name: 'newName' });
       } catch (error) {
         expect(error).toBeInstanceOf(NotFoundException);
         expect(error.message).toBe(
@@ -219,14 +222,14 @@ describe('ShippingCompanyService', () => {
     it('remove should return status and message', async () => {
       const mock = generateShippingCompany();
       const id = mock.id;
-
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'findOne').mockResolvedValue(mock);
       jest
         .spyOn(repository, 'merge')
         .mockReturnValue({ ...mock, isDeleted: true });
       jest.spyOn(repository, 'save').mockResolvedValue(mock);
 
-      const { statusCode, message } = await service.remove(id);
+      const { statusCode, message } = await service.remove(id, userId);
       expect(statusCode).toBe(200);
       expect(message).toEqual(
         `The Shipping Company with id: ${id} has been deleted`,
@@ -235,9 +238,10 @@ describe('ShippingCompanyService', () => {
 
     it('remove should throw NotFoundException if Shipping Company does not exist with Rejects', async () => {
       const id = 1;
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'findOne').mockResolvedValue(null);
 
-      await expect(service.remove(id)).rejects.toThrowError(
+      await expect(service.remove(id, userId)).rejects.toThrowError(
         new NotFoundException(`The Shipping Company with id: ${id} not found`),
       );
     });
